feat!: deploy to Google Cloud Run
BREAKING: The service now deploys to Google Cloud Run (with significant
hoped-for cost savings). Any uses of the service need to be redirected to
the Cloud Run URL. After that has been done, the Google App Engine app
can be disabled, and the unused repostories in the Artifact Repository
can be deleted.